move tests to datastore, consolidate interface
