+++ app/src/main/database/index.test.ts
@@ -2,18 +2,8 @@ import { describe, expect, it, beforeEach, afterEach } from "vitest";
 import fs from "fs";
 import path from "path";
 import os from "os";
-import { MESSAGE_PREVIEW_LENGTH } from "./index";
-import { Datastore } from "../datastore";
-import {
-  FetchStatus,
-  ItemMetadata,
-  PendingEventType,
-  SourceMetadata,
-  Journalist,
-  JournalistMetadata,
-} from "../../types";
+import { DB } from "./index";
 import { Crypto } from "../crypto";
-import { Storage } from "../storage";
 import { setUmask } from "../umask";
 
 describe("Database Component Tests", () => {
@@ -22,7 +12,7 @@ describe("Database Component Tests", () => {
   const testDbPath = path.join(testDbDir, "db.sqlite");
   const originalHomedir = os.homedir;
   let originalUmask: number;
-  let db: Datastore;
+  let db: DB;
   let crypto: Crypto;
 
   beforeAll(() => {
@@ -56,7 +46,7 @@ describe("Database Component Tests", () => {
   });
 
   it("should create database with correct setup", () => {
-    db = new Datastore(crypto, new Storage());
+    db = new DB(crypto);
 
     expect(db).toBeDefined();
     expect(fs.existsSync(testDbPath)).toBe(true);
@@ -66,7 +56,7 @@ describe("Database Component Tests", () => {
   });
 
   it("should handle database closure", () => {
-    db = new Datastore(crypto, new Storage());
+    db = new DB(crypto);
     expect(db).toBeDefined();
 
     db.close();
@@ -74,7 +64,7 @@ describe("Database Component Tests", () => {
   });
 
   it("should create valid database file", () => {
-    db = new Datastore(crypto, new Storage());
+    db = new DB(crypto);
     const stats = fs.statSync(testDbPath);
     expect(stats.isFile()).toBe(true);
     expect(stats.size).toBeGreaterThan(0);
@@ -87,1278 +77,5 @@ describe("Database Component Tests", () => {
     const dirStats = fs.statSync(testDbDir);
     const dirPermissions = dirStats.mode & 0o777;
     expect(dirPermissions).toBe(0o700);
-  });
-});
-
-describe("Database Method Tests", () => {
-  const testHomeDir = path.join(os.tmpdir(), "test-home");
-  const originalHomedir = os.homedir;
-  let db: Datastore;
-  let crypto: Crypto;
-
-  beforeAll(() => {
-    (Crypto as any).instance = undefined;
-    crypto = Crypto.initialize({
-      isQubes: false,
-      submissionKeyFingerprint: "",
-    });
-
-    // Mock crypto.encryptSourceMessage
-    vi.spyOn(crypto, "encryptSourceMessage").mockImplementation(
-      (_plaintext: string, _sourcePublicKey: string): Promise<string> => {
-        return Promise.resolve("test-ciphertext");
-      },
-    );
-  });
-
-  beforeEach(() => {
-    if (fs.existsSync(testHomeDir)) {
-      fs.rmSync(testHomeDir, { recursive: true, force: true });
-    }
-    os.homedir = () => testHomeDir;
-    db = new Datastore(crypto, new Storage());
-  });
-
-  afterEach(() => {
-    if (db) {
-      db.close();
-    }
-    os.homedir = originalHomedir;
-    if (fs.existsSync(testHomeDir)) {
-      fs.rmSync(testHomeDir, { recursive: true, force: true });
-    }
-  });
-
-  function mockSourceMetadata(
-    uuid: string,
-    is_starred?: boolean,
-  ): SourceMetadata {
-    return {
-      uuid: uuid,
-      journalist_designation: "Foo Bar",
-      is_starred: is_starred ?? false,
-      is_seen: false,
-      has_attachment: false,
-      last_updated: "",
-      public_key: "",
-      fingerprint: "",
-    };
-  }
-
-  function mockJournalistMetadata(
-    uuid: string,
-    username?: string,
-  ): JournalistMetadata {
-    return {
-      uuid: uuid,
-      username: username ?? "dellsberg",
-      first_name: null,
-      last_name: null,
-    };
-  }
-
-  function mockItemMetadata(
-    uuid: string,
-    source_uuid: string,
-    kind: "file" | "reply" | "message" = "reply",
-    interaction_count: number = 1,
-  ): ItemMetadata {
-    if (kind === "reply") {
-      return {
-        kind: kind,
-        uuid: uuid,
-        source: source_uuid,
-        size: 50,
-        journalist_uuid: "",
-        is_deleted_by_source: false,
-        seen_by: [],
-        interaction_count: interaction_count,
-      };
-    }
-    return {
-      kind: kind,
-      uuid: uuid,
-      source: source_uuid,
-      size: 50,
-      is_read: false,
-      seen_by: [],
-      interaction_count: interaction_count,
-    };
-  }
-
-  it("getJournalistbyID should return the first matching journalist", () => {
-    // Insert two Journalists
-    db.updateJournalists({
-      journalist1: mockJournalistMetadata("journalist1", "user1"),
-      journalist2: mockJournalistMetadata("journalist2", "user2"),
-    });
-
-    describe("Missing UUID should throw an error", () => {
-      expect(() => db.getJournalistByID("journalist3")).toThrow();
-    });
-
-    let journalist = db.getJournalistByID("journalist1");
-    assertType<Journalist>(journalist);
-    expect(journalist.data.username).toEqual("user1");
-
-    journalist = db.getJournalistByID("journalist2");
-    expect(journalist.data.username).toEqual("user2");
-  });
-
-  it("pending SourceDeleted event should remove source in getSources", () => {
-    // Insert three sources
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-      source2: mockSourceMetadata("source2"),
-      source3: mockSourceMetadata("source3"),
-    });
-
-    let sources = db.getSources();
-    expect(sources.length).toEqual(3);
-    expect(sources.map((s) => s.uuid)).toEqual([
-      "source1",
-      "source2",
-      "source3",
-    ]);
-
-    // Add pending delete for source3
-    db.addPendingSourceEvent("source3", PendingEventType.SourceDeleted);
-
-    // getSources should now return only source1, source2
-    sources = db.getSources();
-    expect(sources.length).toEqual(2);
-    expect(sources.map((s) => s.uuid)).toEqual(["source1", "source2"]);
-  });
-
-  it("pending SourceConversationDeleted should remove all items", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    let sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(3);
-
-    db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.SourceConversationDeleted,
-    );
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(0);
-
-    db.updateSources({
-      source2: mockSourceMetadata("source2"),
-    });
-  });
-
-  it("pending SourceConversationDeleted event should only affect given source", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    let sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(3);
-
-    db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.SourceConversationDeleted,
-    );
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(0);
-
-    // Add source2 with 3 items
-    db.updateSources({
-      source2: mockSourceMetadata("source2"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source2"),
-      item2: mockItemMetadata("item2", "source2"),
-      item3: mockItemMetadata("item3", "source2"),
-    });
-    sourceWithItems = db.getSourceWithItems("source2");
-    expect(sourceWithItems.items.length).toEqual(3);
-  });
-
-  it("pending Starred event should star sources", () => {
-    // Insert three sources
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-      source2: mockSourceMetadata("source2"),
-      source3: mockSourceMetadata("source3"),
-    });
-    let sources = db.getSources();
-    for (const source of sources) {
-      expect(source.data.is_starred).toBe(false);
-    }
-
-    db.addPendingSourceEvent("source1", PendingEventType.Starred);
-    db.addPendingSourceEvent("source2", PendingEventType.Starred);
-
-    sources = db.getSources();
-    for (const source of sources) {
-      if (source.uuid === "source1" || source.uuid === "source2") {
-        expect(source.data.is_starred).toBeTruthy();
-        continue;
-      }
-      expect(source.data.is_starred).toBe(false);
-    }
-  });
-
-  it("pending Unstarred event should unstar source", () => {
-    // Insert three sources
-    db.updateSources({
-      source1: mockSourceMetadata("source1", true),
-      source2: mockSourceMetadata("source2", true),
-      source3: mockSourceMetadata("source3", true),
-    });
-    let sources = db.getSources();
-    for (const source of sources) {
-      expect(source.data.is_starred).toBe(true);
-    }
-
-    db.addPendingSourceEvent("source1", PendingEventType.Unstarred);
-    db.addPendingSourceEvent("source2", PendingEventType.Unstarred);
-
-    sources = db.getSources();
-    for (const source of sources) {
-      if (source.uuid === "source1" || source.uuid === "source2") {
-        expect(source.data.is_starred).toBeFalsy();
-        continue;
-      }
-      expect(source.data.is_starred).toBe(true);
-    }
-  });
-
-  it("multiple Star/Unstarred events should project latest", () => {
-    // Insert three sources
-    db.updateSources({
-      source1: mockSourceMetadata("source1", true),
-      source2: mockSourceMetadata("source2", true),
-      source3: mockSourceMetadata("source3", true),
-    });
-    let sources = db.getSources();
-    for (const source of sources) {
-      expect(source.data.is_starred).toBe(true);
-    }
-
-    db.addPendingSourceEvent("source1", PendingEventType.Unstarred);
-    db.addPendingSourceEvent("source1", PendingEventType.Starred);
-
-    sources = db.getSources();
-    for (const source of sources) {
-      if (source.uuid === "source1") {
-        expect(source.data.is_starred).toBeTruthy();
-        continue;
-      }
-      expect(source.data.is_starred).toBe(true);
-    }
-
-    db.addPendingSourceEvent("source1", PendingEventType.Unstarred);
-    sources = db.getSources();
-    for (const source of sources) {
-      if (source.uuid === "source1") {
-        expect(source.data.is_starred).toBeFalsy();
-        continue;
-      }
-      expect(source.data.is_starred).toBe(true);
-    }
-  });
-
-  it("pending Seen event should mark seen", () => {
-    // Insert three sources
-    db.updateSources({
-      source1: mockSourceMetadata("source1", true),
-      source2: mockSourceMetadata("source2", true),
-      source3: mockSourceMetadata("source3", true),
-    });
-    let sources = db.getSources();
-    for (const source of sources) {
-      expect(source.isRead).toBe(false);
-    }
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    // Partially marking items seen should not update source Seen state
-    db.addPendingItemsSeenBatch("source1", ["item1", "item2"]);
-    for (const source of sources) {
-      expect(source.isRead).toBe(false);
-    }
-
-    // Marking all items as seen should update Seen state
-    db.addPendingItemsSeenBatch("source1", ["item3"]);
-    sources = db.getSources();
-    for (const source of sources) {
-      if (source.uuid === "source1") {
-        expect(source.isRead).toBe(true);
-        continue;
-      }
-      expect(source.isRead).toBe(false);
-    }
-  });
-
-  it("pending ReplySent event should add reply", async () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    let sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(3);
-
-    await db.addPendingReplySentEvent("here is a reply", "source1", 4);
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(4);
-    const reply = sourceWithItems.items[3];
-    expect(reply?.plaintext).toBe("here is a reply");
-  });
-
-  it("getSourceWithItems returns items sorted by interaction_count", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item3: {
-        ...mockItemMetadata("item3", "source1"),
-        interaction_count: 3,
-      },
-      item1: {
-        ...mockItemMetadata("item1", "source1"),
-        interaction_count: 1,
-      },
-      item5: {
-        ...mockItemMetadata("item5", "source1"),
-        interaction_count: 5,
-      },
-      item2: {
-        ...mockItemMetadata("item2", "source1"),
-        interaction_count: 2,
-      },
-      item4: {
-        ...mockItemMetadata("item4", "source1"),
-        interaction_count: 4,
-      },
-    });
-
-    const sourceWithItems = db.getSourceWithItems("source1");
-    const items = sourceWithItems.items;
-
-    // Clone and sort by interaction_count
-    const sortedItems = [...items].sort(
-      (a, b) => a.data.interaction_count - b.data.interaction_count,
-    );
-
-    // Verify the returned items are already sorted
-    expect(items).toEqual(sortedItems);
-  });
-
-  it("pending ItemDeleted event should delete reply", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    let sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(3);
-
-    db.addPendingItemEvent("item1", PendingEventType.ItemDeleted);
-
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(2);
-  });
-
-  it("pending ReplySent and pending SourceConversationDeleted should delete all items in conversation", async () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-    });
-
-    await db.addPendingReplySentEvent("reply text", "source1", 3);
-    let sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(3);
-
-    db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.SourceConversationDeleted,
-    );
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(0);
-  });
-
-  it("pending SourceConversationDeleted and ReplySent should show reply", async () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-    });
-
-    const snowflake1 = db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.SourceConversationDeleted,
-    );
-    let sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(0);
-
-    const snowflake2 = await db.addPendingReplySentEvent(
-      "reply text",
-      "source1",
-      1,
-    );
-    expect(snowflake2 > snowflake1);
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(1);
-    expect(sourceWithItems.items[0].uuid).toMatch(
-      /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i,
-    );
-    expect(sourceWithItems.items[0].plaintext).toBe("reply text");
-  });
-
-  it("pending ReplySent and then pending SourceDeleted should delete source", async () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-    });
-
-    await db.addPendingReplySentEvent("reply text", "source1", 1);
-    let sources = db.getSources();
-    expect(sources.length).toEqual(1);
-
-    db.addPendingSourceEvent("source1", PendingEventType.SourceDeleted);
-    sources = db.getSources();
-    expect(sources.length).toEqual(0);
-  });
-
-  it("pending ReplySent and then pending ItemDeleted should delete reply", async () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-    });
-
-    await db.addPendingReplySentEvent("reply text", "source1", 2);
-    let sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(2);
-    const pendingReplyUuid = sourceWithItems.items[1].uuid;
-
-    // Like `updatePendingEvents()`, insert the reply into `items` to satisfy
-    // `pending_event.item_uuid` → `items.uuid`.
-    db.updateItems({
-      [pendingReplyUuid]: mockItemMetadata(
-        pendingReplyUuid,
-        "source1",
-        "reply",
-        2,
-      ),
-    });
-
-    db.addPendingItemEvent(pendingReplyUuid, PendingEventType.ItemDeleted);
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(1);
-    expect(
-      sourceWithItems.items.find((i) => i.uuid === pendingReplyUuid),
-    ).toBeUndefined();
-  });
-
-  it("pending multiple ItemDeleted events should delete all specified items", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    db.addPendingItemEvent("item1", PendingEventType.ItemDeleted);
-    db.addPendingItemEvent("item3", PendingEventType.ItemDeleted);
-
-    const sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toEqual(1);
-    expect(sourceWithItems.items[0].uuid).toBe("item2");
-  });
-
-  it("addPendingItemsSeenBatch should create Seen events for all items in source", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    const snowflakeIds = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item2",
-      "item3",
-    ]);
-
-    // Should create 3 events for 3 items
-    expect(snowflakeIds.length).toEqual(3);
-
-    // All snowflake IDs should be unique
-    const uniqueIds = new Set(snowflakeIds);
-    expect(uniqueIds.size).toEqual(3);
-  });
-
-  it("addPendingItemsSeenBatch should be idempotent - no duplicates on second call", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    // First call should create 3 events
-    const firstCallIds = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item2",
-      "item3",
-    ]);
-    expect(firstCallIds.length).toEqual(3);
-
-    // Second call should create 0 events (all items already have Seen events)
-    const secondCallIds = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item2",
-      "item3",
-    ]);
-    expect(secondCallIds.length).toEqual(0);
-  });
-
-  it("addPendingItemsSeenBatch should only create events for new items after sync", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-    });
-
-    // First call creates events for 2 items
-    const firstCallIds = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item2",
-    ]);
-    expect(firstCallIds.length).toEqual(2);
-
-    // Add a new item
-    db.updateItems({
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    // Second call should only create event for the new item
-    const secondCallIds = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item2",
-      "item3",
-    ]);
-    expect(secondCallIds.length).toEqual(1);
-  });
-
-  it("addPendingItemsSeenBatch should only affect specified source", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-      source2: mockSourceMetadata("source2"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source2"),
-      item4: mockItemMetadata("item4", "source2"),
-    });
-
-    // Mark source1 items as seen
-    const source1Ids = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item2",
-    ]);
-    expect(source1Ids.length).toEqual(2);
-
-    // source2 items should still be unseen
-    const source2Ids = db.addPendingItemsSeenBatch("source2", [
-      "item3",
-      "item4",
-    ]);
-    expect(source2Ids.length).toEqual(2);
-  });
-
-  it("addPendingItemsSeenBatch should return empty array for source with no items", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    const snowflakeIds = db.addPendingItemsSeenBatch("source1", []);
-    expect(snowflakeIds.length).toEqual(0);
-  });
-
-  it("addPendingItemsSeenBatch should only create events for specified items", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-      item3: mockItemMetadata("item3", "source1"),
-    });
-
-    // Only mark item1 and item3 as seen, not item2
-    const snowflakeIds = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item3",
-    ]);
-
-    // Should only create 2 events
-    expect(snowflakeIds.length).toEqual(2);
-
-    // Calling again with all items should only create event for item2
-    const secondCallIds = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item2",
-      "item3",
-    ]);
-    expect(secondCallIds.length).toEqual(1);
-  });
-
-  it("addPendingItemsSeenBatch should handle empty itemUuids array", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source1"),
-    });
-
-    // Passing empty array should create no events
-    const snowflakeIds = db.addPendingItemsSeenBatch("source1", []);
-    expect(snowflakeIds.length).toEqual(0);
-  });
-
-  it("addPendingItemsSeenBatch should ignore items not in the source", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-      source2: mockSourceMetadata("source2"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source2"),
-    });
-
-    // Try to mark item2 as seen, but it belongs to source2
-    const snowflakeIds = db.addPendingItemsSeenBatch("source1", [
-      "item1",
-      "item2",
-    ]);
-
-    // Should only create event for item1
-    expect(snowflakeIds.length).toEqual(1);
-  });
-
-  it("getPendingEvents should return all pending events with correct structure", async () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-    });
-
-    const snowflakeSource = db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.Starred,
-    );
-    console.log("source Snowflake: ", snowflakeSource);
-    const snowflakeItem = db.addPendingItemEvent(
-      "item1",
-      PendingEventType.ItemDeleted,
-    );
-    const snowflakeReply = await db.addPendingReplySentEvent(
-      "reply text",
-      "source1",
-      1,
-    );
-
-    const events = db.getPendingEvents();
-    expect(events.length).toBe(3);
-    console.log("events: ", events);
-
-    const sourceEvent = events.find((e) => e.id === snowflakeSource);
-    expect(sourceEvent).toBeDefined();
-    expect(sourceEvent!.type).toBe(PendingEventType.Starred);
-    expect(sourceEvent!.target).toHaveProperty("source_uuid", "source1");
-
-    const itemEvent = events.find((e) => e.id === snowflakeItem);
-    expect(itemEvent).toBeDefined();
-    expect(itemEvent!.type).toBe(PendingEventType.ItemDeleted);
-    expect(itemEvent!.target).toHaveProperty("item_uuid", "item1");
-
-    const replyEvent = events.find((e) => e.id === snowflakeReply);
-    expect(replyEvent).toBeDefined();
-    expect(replyEvent!.type).toBe(PendingEventType.ReplySent);
-    expect(replyEvent!.target).toHaveProperty("source_uuid");
-    expect(replyEvent!.data).toHaveProperty("metadata");
-    expect(replyEvent!.data).toHaveProperty("plaintext", "reply text");
-  });
-
-  it("updatePendingEvents should remove successful events from pending_events", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-    });
-
-    // Verify source and item exist
-    const sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems).toBeDefined();
-    expect(sourceWithItems.uuid).toEqual("source1");
-    expect(sourceWithItems.items.length).toBe(1);
-    expect(sourceWithItems.items[0].uuid).toBe("item1");
-
-    const snowflakeSource = db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.Starred,
-    );
-    const snowflakeItem = db.addPendingItemEvent(
-      "item1",
-      PendingEventType.ItemDeleted,
-    );
-
-    // Before update, both events are present
-    let events = db.getPendingEvents();
-    expect(events.length).toBe(2);
-
-    // Simulate server response: both events succeeded (HTTP 200)
-    db.updatePendingEvents({
-      [snowflakeSource.toString()]: [200, ""],
-      [snowflakeItem.toString()]: [200, ""],
-    });
-
-    // After update, no pending events should remain
-    events = db.getPendingEvents();
-    expect(events.length).toBe(0);
-  });
-
-  it("updatePendingEvents should retain failed events", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-    });
-
-    const snowflakeSource = db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.Starred,
-    );
-    const snowflakeItem = db.addPendingItemEvent(
-      "item1",
-      PendingEventType.ItemDeleted,
-    );
-
-    // Simulate server response: only one event succeeded
-    db.updatePendingEvents({
-      [snowflakeSource.toString()]: [200, ""],
-      [snowflakeItem.toString()]: [500, "error"], // failed
-    });
-
-    const events = db.getPendingEvents();
-    expect(events.length).toBe(1);
-    expect(events[0].id).toBe(snowflakeItem);
-    expect(events[0].type).toBe(PendingEventType.ItemDeleted);
-  });
-
-  it("updateItems should upsert items with conflicting uuid", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    // Insert initial item
-    db.updateItems({
-      item1: {
-        ...mockItemMetadata("item1", "source1"),
-        size: 50,
-        interaction_count: 1,
-      },
-    });
-
-    let sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toBe(1);
-    expect(sourceWithItems.items[0].uuid).toBe("item1");
-    expect(sourceWithItems.items[0].data.size).toBe(50);
-    expect(sourceWithItems.items[0].data.interaction_count).toBe(1);
-    expect(sourceWithItems.items[0].fetch_status).toBe(FetchStatus.Initial);
-
-    db.completePlaintextItem("item1", "plaintext");
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toBe(1);
-    expect(sourceWithItems.items[0].uuid).toBe("item1");
-    expect(sourceWithItems.items[0].plaintext).toBe("plaintext");
-    expect(sourceWithItems.items[0].fetch_status).toBe(FetchStatus.Complete);
-
-    // Upsert with same uuid but different data
-    db.updateItems({
-      item1: {
-        ...mockItemMetadata("item1", "source1"),
-        size: 99,
-        interaction_count: 42,
-      },
-    });
-
-    sourceWithItems = db.getSourceWithItems("source1");
-    expect(sourceWithItems.items.length).toBe(1);
-    expect(sourceWithItems.items[0].uuid).toBe("item1");
-    expect(sourceWithItems.items[0].data.size).toBe(99);
-    expect(sourceWithItems.items[0].data.interaction_count).toBe(42);
-    expect(sourceWithItems.items[0].plaintext).toBe("plaintext");
-    expect(sourceWithItems.items[0].fetch_status).toBe(FetchStatus.Complete);
-  });
-
-  it("getSource should return correct message preview", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    let source = db.getSource("source1");
-    expect(source).not.toBeNull();
-    if (source) {
-      expect(source.messagePreview).toBeNull();
-    }
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1", "reply", 1),
-      item2: mockItemMetadata("item2", "source1", "message", 2),
-      item3: mockItemMetadata("item3", "source1", "file", 3),
-    });
-
-    // Set plaintext and filename for items
-    db.completePlaintextItem("item1", "reply message 1");
-    db.completePlaintextItem("item2", "submission message");
-    db.completeFileItem("item3", "/tmp/file3.txt", 1024);
-
-    source = db.getSource("source1");
-    expect(source).not.toBeNull();
-    if (source) {
-      expect(source.messagePreview).not.toBeNull();
-      expect(source.messagePreview?.kind).toBe("file");
-      expect(source.messagePreview?.plaintext).toBe("file3.txt");
-    }
-
-    db.updateItems({
-      item4: mockItemMetadata("item4", "source1", "message", 4),
-    });
-
-    source = db.getSource("source1");
-    expect(source).not.toBeNull();
-    if (source) {
-      expect(source.messagePreview).not.toBeNull();
-      expect(source.messagePreview!.kind).toBe("message");
-      expect(source.messagePreview?.plaintext).toBeNull();
-    }
-
-    db.completePlaintextItem("item4", "latest message");
-    source = db.getSource("source1");
-    expect(source).not.toBeNull();
-    if (source) {
-      expect(source.messagePreview).not.toBeNull();
-      expect(source.messagePreview!.kind).toBe("message");
-      expect(source.messagePreview?.plaintext).toBe("latest message");
-    }
-  });
-
-  it("getSources should return sources with correct last_message_kind, last_message_plaintext, and last_message_filename", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-      source2: mockSourceMetadata("source2"),
-      source3: mockSourceMetadata("source3"),
-      source4: mockSourceMetadata("source4"),
-    });
-
-    db.updateItems({
-      // source1 items
-      item1: mockItemMetadata("item1", "source1", "reply", 1),
-      item2: mockItemMetadata("item2", "source1", "message", 2),
-      // source2 items
-      item3: mockItemMetadata("item3", "source2", "reply", 1),
-      item4: mockItemMetadata("item4", "source2", "file", 2),
-      // source3 items
-      item5: mockItemMetadata("item5", "source3", "message", 1),
-      item6: mockItemMetadata("item6", "source3", "message", 2),
-    });
-
-    db.completePlaintextItem("item1", "reply message 1");
-    db.completePlaintextItem("item2", "message 2");
-    db.completePlaintextItem("item3", "message 1");
-    db.completeFileItem("item4", "/tmp/filename.txt", 2048);
-
-    const sources = db.getSources();
-    const s1 = sources.find((s) => s.uuid === "source1");
-    const s2 = sources.find((s) => s.uuid === "source2");
-    const s3 = sources.find((s) => s.uuid === "source3");
-    const s4 = sources.find((s) => s.uuid === "source4");
-
-    expect(s1).not.toBeNull();
-    if (s1) {
-      expect(s1.messagePreview).not.toBeNull();
-      expect(s1.messagePreview?.kind).toBe("message");
-      expect(s1.messagePreview?.plaintext).toBe("message 2");
-    }
-
-    expect(s2).not.toBeNull();
-    if (s2) {
-      expect(s2.messagePreview).not.toBeNull();
-      expect(s2.messagePreview?.kind).toBe("file");
-      expect(s2.messagePreview?.plaintext).toBe("filename.txt");
-    }
-
-    expect(s3).not.toBeNull();
-    if (s3) {
-      expect(s3.messagePreview).not.toBeNull();
-      expect(s3.messagePreview?.kind).toBe("message");
-      expect(s3.messagePreview?.plaintext).toBeNull();
-    }
-
-    expect(s4).not.toBeNull();
-    if (s4) {
-      expect(s4.messagePreview).toBeNull();
-    }
-  });
-
-  it("messagePreview.plaintext should be truncated to 200 Unicode code points for long messages", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1", "message", 1),
-    });
-
-    // Create a long ASCII-only message (no multi-byte chars) to test basic truncation
-    const longAsciiMessage = "a".repeat(500);
-    db.completePlaintextItem("item1", longAsciiMessage);
-
-    const sources = db.getSources();
-    const source = sources.find((s) => s.uuid === "source1");
-    expect(source).not.toBeNull();
-    expect(source!.messagePreview).not.toBeNull();
-    expect(source!.messagePreview!.plaintext).not.toBeNull();
-    expect(source!.messagePreview!.plaintext!.length).toBe(
-      MESSAGE_PREVIEW_LENGTH,
-    );
-    expect(source!.messagePreview!.plaintext).toBe(
-      "a".repeat(MESSAGE_PREVIEW_LENGTH),
-    );
-
-    // Test getSource (single source query)
-    const singleSource = db.getSource("source1");
-    expect(singleSource).not.toBeNull();
-    expect(singleSource!.messagePreview!.plaintext!.length).toBe(
-      MESSAGE_PREVIEW_LENGTH,
-    );
-  });
-
-  it("messagePreview.plaintext truncation should handle Unicode correctly", () => {
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1", "message", 1),
-    });
-
-    // Create a message with emojis - each emoji is 1 Unicode code point but 2 JS chars
-    // "👋" is 1 code point, so 200 of them = 200 code points = 400 JS chars
-    const emojiMessage = "👋".repeat(300);
-    db.completePlaintextItem("item1", emojiMessage);
-
-    const sources = db.getSources();
-    const source = sources.find((s) => s.uuid === "source1");
-    expect(source).not.toBeNull();
-    expect(source!.messagePreview).not.toBeNull();
-    expect(source!.messagePreview!.plaintext).not.toBeNull();
-    // SQLite truncates at 200 code points; each emoji = 2 JS chars, so 400 JS length
-    expect(source!.messagePreview!.plaintext!.length).toBe(
-      MESSAGE_PREVIEW_LENGTH * 2,
-    );
-    expect(source!.messagePreview!.plaintext).toBe(
-      "👋".repeat(MESSAGE_PREVIEW_LENGTH),
-    );
-  });
-
-  it("deleting a source should cascade to its pending events", () => {
-    // Create a source and add a pending event for it
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-
-    const snowflakeId = db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.Starred,
-    );
-
-    // Verify the event exists
-    let events = db.getPendingEvents();
-    expect(events.length).toBe(1);
-    expect(events[0].id).toBe(snowflakeId);
-
-    // Delete the source (should cascade to its pending events)
-    db.deleteSources(["source1"]);
-
-    // Pending events for the deleted source should be gone
-    events = db.getPendingEvents();
-    expect(events.length).toBe(0);
-  });
-
-  it("deleting an item should cascade to its pending events", () => {
-    // Create a source and item, then add a pending event for the item
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-    });
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-    });
-
-    const snowflakeId = db.addPendingItemEvent(
-      "item1",
-      PendingEventType.ItemDeleted,
-    );
-
-    // Verify the event exists
-    let events = db.getPendingEvents();
-    expect(events.length).toBe(1);
-    expect(events[0].id).toBe(snowflakeId);
-
-    // Delete the item (should cascade to its pending events)
-    db.deleteItems(["item1"]);
-
-    // Pending events for the deleted item should be gone
-    events = db.getPendingEvents();
-    expect(events.length).toBe(0);
-  });
-
-  it("cascading deletion should preserve unrelated pending events", () => {
-    // Create two sources with items
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-      source2: mockSourceMetadata("source2"),
-    });
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source2"),
-    });
-
-    // Add pending events for both
-    const snowflakeSource1 = db.addPendingSourceEvent(
-      "source1",
-      PendingEventType.Starred,
-    );
-    const snowflakeSource2 = db.addPendingSourceEvent(
-      "source2",
-      PendingEventType.Starred,
-    );
-    const snowflakeItem1 = db.addPendingItemEvent(
-      "item1",
-      PendingEventType.Seen,
-    );
-    const snowflakeItem2 = db.addPendingItemEvent(
-      "item2",
-      PendingEventType.Seen,
-    );
-
-    // Verify all events exist
-    let events = db.getPendingEvents();
-    expect(events.length).toBe(4);
-
-    // Delete source1 (should cascade to source1 and item1 events)
-    db.deleteSources(["source1"]);
-
-    // Only events for source2/item2 should remain
-    events = db.getPendingEvents();
-    expect(events.length).toBe(2);
-
-    const remainingIds = events.map((e) => e.id);
-    expect(remainingIds).toContain(snowflakeSource2);
-    expect(remainingIds).toContain(snowflakeItem2);
-    expect(remainingIds).not.toContain(snowflakeSource1);
-    expect(remainingIds).not.toContain(snowflakeItem1);
-  });
-
-  it("cascading deletion should handle multiple sources and events", () => {
-    // Create sources and items
-    db.updateSources({
-      source1: mockSourceMetadata("source1"),
-      source2: mockSourceMetadata("source2"),
-    });
-    db.updateItems({
-      item1: mockItemMetadata("item1", "source1"),
-      item2: mockItemMetadata("item2", "source2"),
-    });
-
-    // Add multiple pending events
-    db.addPendingSourceEvent("source1", PendingEventType.Starred);
-    db.addPendingSourceEvent("source1", PendingEventType.Unstarred);
-    db.addPendingSourceEvent("source2", PendingEventType.Starred);
-    db.addPendingItemEvent("item1", PendingEventType.Seen);
-    db.addPendingItemEvent("item2", PendingEventType.Seen);
-
-    // Verify all 5 events exist
-    let events = db.getPendingEvents();
-    expect(events.length).toBe(5);
-
-    // Delete both sources (should cascade to all related events)
-    db.deleteSources(["source1", "source2"]);
-
-    // All events should be deleted via cascade
-    events = db.getPendingEvents();
-    expect(events.length).toBe(0);
-  });
-
-  describe("updateFetchStatus", () => {
-    it("should reset fetch_progress to 0 when updating to Initial status", () => {
-      db.updateSources({
-        source1: mockSourceMetadata("source1"),
-      });
-      db.updateItems({
-        item1: mockItemMetadata("item1", "source1", "file"),
-      });
-
-      // Simulate a partial download by setting progress
-      db.setDownloadInProgress("item1", 50000);
-
-      // Verify progress was set
-      let item = db.getItem("item1");
-      expect(item?.fetch_progress).toBe(50000);
-
-      // Now update status to Initial (simulating Cancel)
-      db.updateFetchStatus("item1", FetchStatus.Initial);
-
-      // Verify progress was reset to 0
-      item = db.getItem("item1");
-      expect(item?.fetch_status).toBe(FetchStatus.Initial);
-      expect(item?.fetch_progress).toBe(0);
-    });
-
-    it("should reset fetch_progress to 0 when updating to DownloadInProgress status", () => {
-      db.updateSources({
-        source1: mockSourceMetadata("source1"),
-      });
-      db.updateItems({
-        item1: mockItemMetadata("item1", "source1", "file"),
-      });
-
-      // Simulate a failed download with progress
-      db.setDownloadInProgress("item1", 100000);
-      db.failDownload("item1");
-
-      // Verify progress was preserved during failure
-      let item = db.getItem("item1");
-      expect(item?.fetch_progress).toBe(100000);
-      expect(item?.fetch_status).toBe(FetchStatus.FailedDownloadRetryable);
-
-      // Now update status to DownloadInProgress (simulating Retry)
-      db.updateFetchStatus("item1", FetchStatus.DownloadInProgress);
-
-      // Verify progress was reset to 0
-      item = db.getItem("item1");
-      expect(item?.fetch_status).toBe(FetchStatus.DownloadInProgress);
-      expect(item?.fetch_progress).toBe(0);
-    });
-
-    it("should NOT reset fetch_progress when updating to other statuses", () => {
-      db.updateSources({
-        source1: mockSourceMetadata("source1"),
-      });
-      db.updateItems({
-        item1: mockItemMetadata("item1", "source1", "file"),
-      });
-
-      // Set some progress
-      db.setDownloadInProgress("item1", 75000);
-
-      // Update to Paused status
-      db.updateFetchStatus("item1", FetchStatus.Paused);
-
-      // Verify progress was NOT reset
-      const item = db.getItem("item1");
-      expect(item?.fetch_status).toBe(FetchStatus.Paused);
-      expect(item?.fetch_progress).toBe(75000);
-    });
-
-    it("should allow retry after terminal failure when status is reset", () => {
-      db.updateSources({
-        source1: mockSourceMetadata("source1"),
-      });
-      db.updateItems({
-        item1: mockItemMetadata("item1", "source1", "file"),
-      });
-
-      // Simulate a terminal failure with progress
-      db.setDownloadInProgress("item1", 50000);
-      db.terminallyFailItem("item1");
-
-      let item = db.getItem("item1");
-      expect(item?.fetch_status).toBe(FetchStatus.FailedTerminal);
-
-      // Reset to Initial (Cancel) should reset progress
-      db.updateFetchStatus("item1", FetchStatus.Initial);
-
-      item = db.getItem("item1");
-      expect(item?.fetch_status).toBe(FetchStatus.Initial);
-      expect(item?.fetch_progress).toBe(0);
-
-      // Now it can be processed again
-      db.updateFetchStatus("item1", FetchStatus.DownloadInProgress);
-
-      item = db.getItem("item1");
-      expect(item?.fetch_status).toBe(FetchStatus.DownloadInProgress);
-      expect(item?.fetch_progress).toBe(0);
-    });
   });
 });
